+++ ragstack/colbert/__init__.py
@@ -1,13 +1,13 @@
 from .colbert_embedding import ColbertTokenEmbeddings
-from .cassandra_db import AstraDB
-from .astra_retriever import ColbertAstraRetriever, max_similarity_torch
+from .cassandra_db import CassandraDB
+from .cassandra_retriever import ColbertCassandraRetriever, max_similarity_torch
 from .token_embedding import PerTokenEmbeddings, PassageEmbeddings, TokenEmbeddings
 from .constant import DEFAULT_COLBERT_MODEL, DEFAULT_COLBERT_DIM
 
 __all__ = (
     ColbertTokenEmbeddings,
-    AstraDB,
-    ColbertAstraRetriever,
+    CassandraDB,
+    ColbertCassandraRetriever,
     max_similarity_torch,
     PerTokenEmbeddings,
     PassageEmbeddings,
